+++ Editor/Package.cs
@@ -7,10 +7,6 @@ namespace HananokiEditor.CustomProjectBrowser {
     public const string name = "CustomProjectBrowser";
     public const string nameNicify = "Custom Project Browser";
     public const string editorPrefName = "Hananoki.CustomProjectBrowser";
-    public const string version = "0.7.0";
-		[HananokiEditorMDViewerRegister]
-		public static string MDViewerRegister() {
-			return "68cd53ef6b462bc48a6292d6e84fdfa4";
-		}
+    public const string version = "0.8.0";
   }
 }