+++ Editor/Utils.cs
@@ -28,25 +28,5 @@ namespace HananokiEditor.CustomProjectBrowser {
 			EditorApplication.update -= _DeleyAttachDockPane;
 #endif
 		}
-
-
-#if UNITY_2019_1_OR_NEWER
-		static void _DeleyAttachToolbar() {
-			if( !GUIToolbar.Attach() ) return;
-
-			EditorApplication.update -= _DeleyAttachToolbar;
-		}
-
-		public static void AttachToolbar() {
-			EditorApplication.update -= _DeleyAttachToolbar;
-			EditorApplication.update += _DeleyAttachToolbar;
-		}
-
-
-		public static void DetachToolbar() {
-			GUIToolbar.Dettach();
-			EditorApplication.update -= _DeleyAttachToolbar;
-		}
-#endif
 	}
 }